+++ src/main/java/cech12/extendedmushrooms/block/MushroomSlabBlock.java
@@ -1,14 +1,14 @@
 package cech12.extendedmushrooms.block;
 
-import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.block.SlabBlock;
 import net.minecraft.core.Direction;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.BlockGetter;
 
-public class VerticalPlanksBlock extends Block {
+public class MushroomSlabBlock extends SlabBlock {
 
-    public VerticalPlanksBlock(Properties properties) {
+    public MushroomSlabBlock(Properties properties) {
         super(properties);
     }